remove some unnecessary functions, move roles to external file
