+++ package.json
@@ -1,6 +1,7 @@
 {
   "type": "module",
   "dependencies": {
-    "discord.js": "^14.15.3"
+    "discord.js": "^14.15.3",
+    "dotenv": "^16.4.5"
   }
 }
